# docs/swagger.yaml
definitions:
  auth.RefreshInput:
    properties:
      accessToken:
        minLength: 10
        type: string
    required:
    - accessToken
    type: object
  auth.SignInResponse:
    properties:
      auth:
        $ref: '#/definitions/models.AuthData'
      user:
        $ref: '#/definitions/models.UserModel'
    type: object
  auth.SignInput:
    properties:
      password:
        minLength: 10
        type: string
      username:
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
  models.AuthData:
    properties:
      accessExpire:
        type: integer
      accessToken:
        type: string
      expiresIn:
        type: integer
      refreshToken:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.UserModel:
    properties:
      created_at:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/refresh-tokens:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      tags:
      - Auth
  /api/auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SignInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.SignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      tags:
      - Auth
  /api/auth/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SignInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      tags:
      - Auth
swagger: "2.0"